pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vectordash"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-multipart>=0.0.9",
  "pydantic>=2.8",
  "pandas>=2.2",
  "openpyxl>=3.1",
  "numpy>=1.26",
  "jinja2>=3.1",
  "sqlmodel>=0.0.21",
  "sse-starlette>=2.2.1",
  "matplotlib>=3.9",
  "faiss-cpu>=1.8.0",
  "sentence-transformers>=3.0",
  "rank-bm25>=0.2.2",
  "scikit-learn>=1.4",
  "python-docx>=1.1.2",
  "beautifulsoup4>=4.12,<5"
]

[tool.setuptools.packages.find]
# Only package our application code; exclude data/ etc. to avoid auto-discovery errors
include = ["app*"]
exclude = ["data*", "tests*", "dist*", "build*"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E","F","I","UP","B","SIM","PL","RUF"]
